+++ src/build_number.h
@@ -1,1 +1,1 @@
-#define BUILD 591
+#define BUILD 592